backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-docx==1.1.0
pypdf==3.17.4
openai==1.3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
